server/requirements.txt
# Backend Requirements - requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
python-dotenv==1.1.1
PyMySQL==1.1.0